requirements.txt
pandas
geopandas
shapely
networkx
fiona
pyyaml
jinja2
tqdm
pydantic
pytest
black
pyarrow